Move findButtonByText into global helpers to reduce and share code
